+++ src/presentation/controllers/signup/signup.controller.spec.ts
@@ -1,8 +1,6 @@
 import { SignupController } from './signup.controller';
-import { InvalidParamError, MissingParamError, ServerError } from '../errors';
-import { EmailValidator } from '../protocols';
-import { AddAccount, AddAccountModel } from '../../domain/use-cases/add-account';
-import { AccountModel } from '../../domain/models/account';
+import { InvalidParamError, MissingParamError, ServerError } from '../../errors';
+import { EmailValidator, AddAccount, AddAccountModel, AccountModel } from './signup.protocols';
 
 const makeEmailValidator = (): EmailValidator => {
   class EmailValidatorStub implements EmailValidator {
