+++ src/presentation/controllers/signup/signup.controller.ts
@@ -1,7 +1,12 @@
-import { HttpRequest, HttpResponse, EmailValidator, Controller } from '../protocols';
-import { MissingParamError, InvalidParamError } from '../errors';
-import { badRequest, serverError } from '../helpers/http.helper';
-import { AddAccount } from '../../domain/use-cases/add-account';
+import {
+  HttpRequest,
+  HttpResponse,
+  Controller,
+  EmailValidator,
+  AddAccount,
+} from './signup.protocols';
+import { MissingParamError, InvalidParamError } from '../../errors';
+import { badRequest, serverError } from '../../helpers/http.helper';
 
 export class SignupController implements Controller {
   private readonly emailValidator: EmailValidator;
